requirements.txt
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.6.1
fastapi==0.75.2
fastapi-utils==0.2.1
greenlet==3.0.3
h11==0.13.0
idna==3.3
loguru==0.6.0
py-postgresql==1.2.2
pydantic==1.9.0
PyJWT==2.3.0
pymongo==4.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
